egovframe-common-components 5.0.0

+++ src/main/java/egovframework/com/cop/smt/sam/web/EgovAllSchdulManageController.java
@@ -3,8 +3,6 @@ package egovframework.com.cop.smt.sam.web;
 import java.util.List;
 import java.util.Map;
 
-import javax.annotation.Resource;
-
 import org.egovframe.rte.fdl.property.EgovPropertyService;
 import org.egovframe.rte.psl.dataaccess.util.EgovMap;
 import org.egovframe.rte.ptl.mvc.tags.ui.pagination.PaginationInfo;
@@ -18,6 +16,7 @@ import egovframework.com.cmm.ComDefaultVO;
 import egovframework.com.cmm.EgovMessageSource;
 import egovframework.com.cmm.annotation.IncludedInfo;
 import egovframework.com.cop.smt.sam.service.EgovAllSchdulManageService;
+import jakarta.annotation.Resource;
 /**
  * 전체일정을 처리하는 Controller Class 구현
  * @author 공통서비스 장동한
@@ -87,7 +86,7 @@ public class EgovAllSchdulManageController {
         model.addAttribute("searchKeyword", commandMap.get("searchKeyword") == null ? "" : (String)commandMap.get("searchKeyword"));
         model.addAttribute("searchCondition", commandMap.get("searchCondition") == null ? "" : (String)commandMap.get("searchCondition"));
 
-        int totCnt = (Integer)egovAllSchdulManageService.selectAllSchdulManageListCnt(searchVO);
+        int totCnt = egovAllSchdulManageService.selectAllSchdulManageListCnt(searchVO);
 		paginationInfo.setTotalRecordCount(totCnt);
         model.addAttribute("paginationInfo", paginationInfo);
