+++ src/test/java/egovframework/com/cop/bbs/service/impl/EgovArticleDAOTest_selectArticleList.java
@@ -1,6 +1,6 @@
 package egovframework.com.cop.bbs.service.impl;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.Date;
 import java.util.List;
@@ -8,7 +8,7 @@ import java.util.List;
 import org.egovframe.rte.fdl.cmmn.exception.FdlException;
 import org.egovframe.rte.fdl.idgnr.EgovIdGnrService;
 import org.egovframe.rte.fdl.string.EgovDateUtil;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.test.context.ContextConfiguration;
@@ -96,7 +96,7 @@ public class EgovArticleDAOTest_selectArticleList extends EgovTestV1 {
 		Board board = new Board();
 
 		try {
-			board.setNttId(egovNttIdGnrService.getNextIntegerId());
+			board.setNttId((long) egovNttIdGnrService.getNextIntegerId());
 		} catch (FdlException e) {
 			log.error("FdlException egovNttIdGnrService");
 		}
